Insert validation part

+++ Website/patient_delete.php
@@ -1,15 +1,20 @@
 <?php
-$con = mysql_connect("localhost","root","Conestoga1", "pInfo");
-if (!$con) {
-  die('Could not connect: ' . mysql_error());
+if(empty($_GET[id])){
+  echo "invalid input";
 }
+else {
+  $con = mysql_connect("localhost","root","Conestoga1", "pInfo");
+  if (!$con) {
+    die('Could not connect: ' . mysql_error());
+  }
 
-mysql_query("Use pInfo", $con);
-$sql = "DELETE FROM pLog where id=$_GET[id]";
+  mysql_query("Use pInfo", $con);
+  $sql = "DELETE FROM pLog where id=$_GET[id]";
 
-if (!mysql_query($sql,$con)) {
-  die('Error: ' . mysql_error());
+  if (!mysql_query($sql,$con)) {
+    die('Error: ' . mysql_error());
+  }
+  echo "1 record deleted";
+  mysql_close($con);
 }
-echo "1 record deleted";
-mysql_close($con);
 ?>
